+++ NEWS.md
@@ -1,5 +1,7 @@
 # raadtools dev
 
+* NSIDC netcdf files needed to disable 'rescale'. We don't currently control that. 
+
 * New scheme for NSIDC netcdf files.  We now use these directly, there's no use for 'vrt_dsn'. Now depending
 on raadfiles 0.1.3.9041, in raadfiles the dates are culled to be distinct. Here the empty dates (no variable exists in the shell netcdf), we remove those in `icefiles()`.